feat(signers-cryptography): move and fix BDD naming in algorithm tests (ENG-2313)
- Move aes-gcm.test.ts and fpe.test.ts from __tests__/algorithms/ to
  algorithms/symmetric/ (co-located with source files per convention)
- Update imports to reflect new co-located paths
- Fix 3 BDD naming violations: "should X" → active present-tense verbs
